cozy_chess

Module util

Additional common utilities

Functions

display_san_move

Returns an object that allows printing a Move in Standard Algebraic Notation.

display_uci_move

Returns an object that allows printing a Move in UCI format.

parse_san_move

Parses a Standard Algebraic Notation move into a Move.

parse_uci_move

Parses a UCI move into a Move.